feat: consolidate providers into a single crate (#36)
