+++ Cargo.toml
@@ -6,9 +6,7 @@ members = [
     "authkestra-session-memory",
     "authkestra-session-redis",
     "authkestra-session-sql",
-    "authkestra-providers-github",
-    "authkestra-providers-google",
-    "authkestra-providers-discord",
+    "authkestra-providers",
     "authkestra-axum",
     "authkestra-macros",
     "authkestra-oidc",
