--- authkestra-providers-discord/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "authkestra-providers-discord"
-version = "0.1.1"
-edition = "2021"
-description = "Discord OAuth provider for the authkestra framework"
-repository = "https://github.com/marcjazz/authkestra"
-license = "MIT OR Apache-2.0"
-keywords = ["auth", "oauth2", "discord"]
-categories = ["authentication"]
-
-[dependencies]
-authkestra-engine = { version = "0.1.1", path = "../authkestra-engine" }
-async-trait = "0.1"
-reqwest = { version = "0.13.1", features = ["json", "form"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-urlencoding = "2.1"
-
-[dev-dependencies]
-wiremock = "0.6.5"
-tokio = { version = "1.0", features = ["full"] }
