+++ authkestra-providers/Cargo.toml
@@ -1,12 +1,18 @@
 [package]
-name = "authkestra-providers-github"
+name = "authkestra-providers"
 version = "0.1.1"
 edition = "2021"
-description = "GitHub OAuth provider for the authkestra framework"
+description = "Consolidated OAuth providers for the authkestra framework"
 repository = "https://github.com/marcjazz/authkestra"
 license = "MIT OR Apache-2.0"
-keywords = ["auth", "oauth2", "github"]
+keywords = ["auth", "oauth2", "github", "google", "discord"]
 categories = ["authentication"]
+
+[features]
+default = []
+github = []
+google = []
+discord = ["urlencoding"]
 
 [dependencies]
 authkestra-engine = { version = "0.1.1", path = "../authkestra-engine" }
@@ -14,7 +20,8 @@ async-trait = "0.1"
 reqwest = { version = "0.13.1", features = ["json", "form"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+urlencoding = { version = "2.1", optional = true }
 
 [dev-dependencies]
 wiremock = "0.6.5"
-tokio = { version = "1.0", features = ["macros", "rt-multi-thread"] }
+tokio = { version = "1.0", features = ["full"] }
