+++ authkestra-providers/src/lib.rs
@@ -1,0 +1,8 @@
+#[cfg(feature = "github")]
+pub mod github;
+
+#[cfg(feature = "google")]
+pub mod google;
+
+#[cfg(feature = "discord")]
+pub mod discord;
