+++ authkestra-providers/tests/discord_oauth.rs
@@ -2,66 +2,67 @@ use authkestra_engine::{
     state::{Identity, OAuthToken},
     OAuthProvider,
 };
-use authkestra_providers_github::GithubProvider;
+use authkestra_providers::discord::DiscordProvider;
 use wiremock::matchers::{body_string_contains, header, method, path};
 use wiremock::{Mock, MockServer, ResponseTemplate};
 
 #[tokio::test]
-async fn test_github_oauth_flow() {
+async fn test_discord_oauth_flow() {
     // Start a mock server
     let server = MockServer::start().await;
 
-    // Mock the GitHub token endpoint
+    // Mock the Discord token endpoint
     Mock::given(method("POST"))
-        .and(path("/login/oauth/access_token"))
-        .and(header("Accept", "application/json"))
+        .and(path("/api/oauth2/token"))
         .and(body_string_contains("code=test_code"))
         .respond_with(
             ResponseTemplate::new(200)
                 .append_header("content-type", "application/json")
                 .set_body_json(serde_json::json!({
                     "access_token": "test_access_token",
-                    "token_type": "bearer"
+                    "token_type": "Bearer",
+                    "expires_in": 604800,
+                    "refresh_token": "test_refresh_token",
+                    "scope": "identify email"
                 })),
         )
         .mount(&server)
         .await;
 
-    // Mock the GitHub user info endpoint
+    // Mock the Discord user info endpoint
     Mock::given(method("GET"))
-        .and(path("/user"))
+        .and(path("/api/users/@me"))
         .and(header("Authorization", "Bearer test_access_token"))
         .respond_with(
             ResponseTemplate::new(200)
                 .append_header("content-type", "application/json")
                 .set_body_json(serde_json::json!({
-                    "id": 123,
-                    "login": "test_user",
-                    "email": "test@example.com"
+                    "id": "123456789",
+                    "username": "testuser",
+                    "discriminator": "0001",
+                    "email": "test@example.com",
+                    "verified": true
                 })),
         )
         .mount(&server)
         .await;
 
-    let provider = GithubProvider::new(
+    let provider = DiscordProvider::new(
         "test_client_id".to_string(),
         "test_client_secret".to_string(),
         format!("{}/callback", server.uri()),
     )
     .with_test_urls(
-        format!("{}/login/oauth/authorize", server.uri()),
-        format!("{}/login/oauth/access_token", server.uri()),
-        format!("{}/user", server.uri()),
+        format!("{}/api/oauth2/token", server.uri()),
+        format!("{}/api/users/@me", server.uri()),
+        format!("{}/api/oauth2/token/revoke", server.uri()),
     );
 
     // Simulate the authorization URL generation
-    let authorize_url = provider.get_authorization_url("test_state", &["user:email"], None);
-    assert!(authorize_url.starts_with(&format!("{}/login/oauth/authorize", server.uri())));
+    let authorize_url = provider.get_authorization_url("test_state", &["identify", "email"], None);
     assert!(authorize_url.contains("state=test_state"));
+    assert!(authorize_url.contains("client_id=test_client_id"));
 
-    // In a real scenario, the user would be redirected to GitHub, authorize the app,
-    // and then be redirected back to the callback URL with a code.
-    // For testing, we'll directly use the mocked code.
     let code = "test_code";
 
     let (identity, token_response): (Identity, OAuthToken) = provider
@@ -70,7 +71,12 @@ async fn test_github_oauth_flow() {
         .expect("Failed to exchange code");
 
     assert_eq!(token_response.access_token, "test_access_token".to_string());
-    assert_eq!(identity.external_id, "123");
-    assert_eq!(identity.username, Some("test_user".to_string()));
+    assert_eq!(
+        token_response.refresh_token,
+        Some("test_refresh_token".to_string())
+    );
+
+    assert_eq!(identity.external_id, "123456789");
+    assert_eq!(identity.username, Some("testuser#0001".to_string()));
     assert_eq!(identity.email, Some("test@example.com".to_string()));
 }
