+++ authkestra/Cargo.toml
@@ -14,9 +14,7 @@ authkestra-session = { version = "0.1.1", path = "../authkestra-session", option
 authkestra-oidc = { version = "0.1.2", path = "../authkestra-oidc", optional = true }
 authkestra-axum = { version = "0.1.2", path = "../authkestra-axum", optional = true }
 authkestra-actix = { version = "0.1.2", path = "../authkestra-actix", optional = true }
-authkestra-providers-github = { version = "0.1.1", path = "../authkestra-providers-github", optional = true }
-authkestra-providers-google = { version = "0.1.1", path = "../authkestra-providers-google", optional = true }
-authkestra-providers-discord = { version = "0.1.1", path = "../authkestra-providers-discord", optional = true }
+authkestra-providers = { version = "0.1.1", path = "../authkestra-providers", optional = true }
 
 [dev-dependencies]
 tokio = { version = "1", features = ["full"] }
@@ -37,6 +35,7 @@ axum = ["dep:authkestra-axum", "authkestra-axum/flow", "authkestra-axum/session"
 actix = ["dep:authkestra-actix", "authkestra-actix/flow", "authkestra-actix/session", "authkestra-actix/token", "authkestra-actix/guard"]
 
 # Providers
-github = ["dep:authkestra-providers-github"]
-google = ["dep:authkestra-providers-google"]
-discord = ["dep:authkestra-providers-discord"]
+github = ["authkestra-providers/github"]
+google = ["authkestra-providers/google"]
+discord = ["authkestra-providers/discord"]
+
